+++ clang/lib/Serialization/ASTReaderDecl.cpp
@@ -2293,6 +2293,8 @@ void ASTDeclReader::VisitCXXDeductionGuideDecl(CXXDeductionGuideDecl *D) {
   VisitFunctionDecl(D);
   D->setDeductionCandidateKind(
       static_cast<DeductionCandidate>(Record.readInt()));
+  D->SourceDeductionGuide.setPointer(readDeclAs<CXXDeductionGuideDecl>());
+  D->setGeneratedFromInheritedConstructor(Record.readBool());
 }
 
 void ASTDeclReader::VisitCXXMethodDecl(CXXMethodDecl *D) {