+++ clang/lib/Serialization/ASTWriterDecl.cpp
@@ -790,6 +790,8 @@ void ASTDeclWriter::VisitCXXDeductionGuideDecl(CXXDeductionGuideDecl *D) {
   Record.AddDeclRef(D->Ctor);
   VisitFunctionDecl(D);
   Record.push_back(static_cast<unsigned char>(D->getDeductionCandidateKind()));
+  Record.AddDeclRef(D->SourceDeductionGuide.getPointer());
+  Record.push_back(D->isGeneratedFromInheritedConstructor());
   Code = serialization::DECL_CXX_DEDUCTION_GUIDE;
 }